refuse a signal number no signal has, where it arrives
`kill` took the number out of the register and passed it on unchecked. What
read it masked it into range on the way to the pending set, and what compared
it read it whole, so the two disagreed:

    kill SIGSTOP = 0
    kill 73 = 0
    waitpid = 0, status 0       still alive, with the kill bit pending
    kill SIGKILL = 0
    waitpid = 5, status 0x9

73 masked to six bits is 9, so the send set the bit for the kill signal; the
code that restarts a stopped task for a kill compared 73 against 9 and left it
stopped, holding a kill it could not be rid of and would take the moment it was
continued. Negative numbers fold into the range the same way.

A signal is now a type whose one constructor says no to a number no signal has,
and which carries its bit and its table index as methods. Every mask and every
cast between a number and either of those is gone, and `kill` and `rt_sigaction`
answer EINVAL at the register.

Two casts are left, in files this change does not own: the mask in each
architecture's `enter_signal_handler`, which now only ever sees a number the
boundary already checked.

+++ kernel/src/syscall/mod.rs
@@ -266,7 +266,7 @@ fn handle(number: u64, args: &[u64; 6], frame: &mut TrapFrame) -> SysResult {
         }
 
         // ---- signals (default actions only) ---------------------------
-        nr::RT_SIGACTION => proc::rt_sigaction(args[0] as usize, args[1], args[2]),
+        nr::RT_SIGACTION => proc::rt_sigaction(args[0] as i32, args[1], args[2]),
         nr::RT_SIGPROCMASK => proc::rt_sigprocmask(args[0] as u32, args[1], args[2]),
         nr::RT_SIGSUSPEND => Err(Errno::EINTR),
         nr::SIGALTSTACK => Ok(0),
